Refactor Header component to use custom wallet hooks, removing direct wallet selector implementation and simplifying wallet connection logic.
